Improve responses (part 1)
1. app/endpoints/fabric.py

- build_response()

New method to build client responses.

- build_nv_pairs

New method to construct the nvPairs object in client responses.

2. app/models/fabric.py

- FabricResponseModel()

New class which determines the structure of the response sent to clients.  For now, we've added id (as a uuid.UUID) and nvPairs.

- NvPairs()

New class which determines the structure of the nvPairs portion of a fabric response.

- FabricBase()

Alphabetize properties

- Fabric()

Modify the docstring for clarity

- FabricPublic()

Remove unused class.