Handle case where fabric exists more gracefully
1. app/v1/endpoints/fabric.py

In post_fabric(), when we POST the same fabric twice, we were getting a 500 error the second time because the SQLlite Integrity error object is not JSON serializable:

TypeError: Object of type IntegrityError is not JSON serializable

Rather than return the error object, we return what NDFC returns:

"ND Site with name <fabric_name> already exists."

Also changed the status_code from 400 to 500, which is what NDFC returns.

Finally, we now also call session.rollback() in the exception handler.