Update integration tests
1. tests/integration/fabric.yaml

- Rename fabric from f1 to TEST_FABRIC
- Modify asserts for int and bool parameters to compare to a list of unquoted and quoted values

ND 3.2.1e returns all values as quoted strings whereas ND 4.0 returns values in their native formats as defined in the schema.  For the integration tests to work against both versions, we've changed the asserts to allow both.

2. tests/integration/inventory/group_vars/ndfc/params.yaml

- Remove unused ansible_port and remote_port
- Add two sections at the bottom for ansible_httpapi_port.  One for running against a container, and the other for running against fastapi outside of a container.

By default, the container config is uncommented.  Change this if you want to run the integration tests against fastapi outside of a container.