discover_post: Update ccStatus, operStatus, and fabricId
For now, let's update a switch's parameters that have to do with status, to match what we are storing in the overview table.  This way, we can reference the switch params when updating the overview table.  This isn't NDFC behavior, but may be good enough, at least for now.  We can revisit later.

Another approach, which we'll look into, would be to ignore ccStatus, operStatus, switchRole, etc, when updating the overview table if these are null ("") in the switch record.

1. app/v1/endpoints/lan_fabric/rest/control/fabrics/inventory/discover_post.py

- On POST, update switch parameters to match what is stored in the overview tables.